prod_requirements.lock.txt
annotated-doc==0.0.4 ; python_version >= "3.12" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.13.0 ; python_version >= "3.12" and python_version < "4.0"
argh==0.31.3 ; python_version >= "3.12" and python_version < "4.0"
argon2-cffi-bindings==25.1.0 ; python_version >= "3.12" and python_version < "4.0"
argon2-cffi==25.1.0 ; python_version >= "3.12" and python_version < "4.0"
babel==2.18.0 ; python_version >= "3.12" and python_version < "4.0"
cairocffi==1.7.1 ; python_version >= "3.12" and python_version < "4.0"
cairosvg==2.9.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2025.10.5 ; python_version >= "3.12" and python_version < "4.0"
cffi==2.0.0 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.4.6 ; python_version >= "3.12" and python_version < "4.0"
click==8.3.0 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and (platform_system == "Windows" or sys_platform == "win32")
cryptography==46.0.5 ; python_version >= "3.12" and python_version < "4.0"
cssselect2==0.9.0 ; python_version >= "3.12" and python_version < "4.0"
defusedxml==0.7.1 ; python_version >= "3.12" and python_version < "4.0"
dnspython==2.8.0 ; python_version >= "3.12" and python_version < "4.0"
email-validator==2.3.0 ; python_version >= "3.12" and python_version < "4.0"
et-xmlfile==2.0.0 ; python_version >= "3.12" and python_version < "4.0"
fastapi==0.135.2 ; python_version >= "3.12" and python_version < "4.0"
h11==0.16.0 ; python_version >= "3.12" and python_version < "4.0"
httpcore==1.0.9 ; python_version >= "3.12" and python_version < "4.0"
httptools==0.7.1 ; python_version >= "3.12" and python_version < "4.0"
httpx==0.28.1 ; python_version >= "3.12" and python_version < "4.0"
idna==3.11 ; python_version >= "3.12" and python_version < "4.0"
isodate==0.7.2 ; python_version >= "3.12" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.12" and python_version < "4.0"
jinja2==3.1.6 ; python_version >= "3.12" and python_version < "4.0"
lxml==6.0.2 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==3.0.3 ; python_version >= "3.12" and python_version < "4.0"
msoffcrypto-tool==5.4.2 ; python_version >= "3.12" and python_version < "4.0"
olefile==0.47 ; python_version >= "3.12" and python_version < "4.0"
openpyxl==3.1.5 ; python_version >= "3.12" and python_version < "4.0"
pillow==12.1.1 ; python_version >= "3.12" and python_version < "4.0"
psycopg-binary==3.3.3 ; python_version >= "3.12" and python_version < "4.0" and implementation_name != "pypy"
psycopg-pool==3.3.0 ; python_version >= "3.12" and python_version < "4.0"
psycopg==3.3.3 ; python_version >= "3.12" and python_version < "4.0"
pycparser==2.23 ; python_version >= "3.12" and python_version < "4.0" and implementation_name != "PyPy"
pydantic-core==2.41.5 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.12.5 ; python_version >= "3.12" and python_version < "4.0"
pymemcache==4.0.0 ; python_version >= "3.12" and python_version < "4.0"
pyotp==2.9.0 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.1.1 ; python_version >= "3.12" and python_version < "4.0"
python-http-client==3.3.7 ; python_version >= "3.12" and python_version < "4.0"
python-multipart==0.0.22 ; python_version >= "3.12" and python_version < "4.0"
python3-saml==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
pyyaml==6.0.3 ; python_version >= "3.12" and python_version < "4.0"
qrcode==8.2 ; python_version >= "3.12" and python_version < "4.0"
requests==2.33.0 ; python_version >= "3.12" and python_version < "4.0"
resend==2.26.0 ; python_version >= "3.12" and python_version < "4.0"
sendgrid==6.12.5 ; python_version >= "3.12" and python_version < "4.0"
starlette==0.50.0 ; python_version >= "3.12" and python_version < "4.0"
tinycss2==1.5.1 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.15.0 ; python_version >= "3.12" and python_version < "4.0"
typing-inspection==0.4.2 ; python_version >= "3.12" and python_version < "4.0"
tzdata==2025.2 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "win32"
ua-parser-builtins==202601 ; python_version >= "3.12" and python_version < "4.0"
ua-parser==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.6.3 ; python_version >= "3.12" and python_version < "4.0"
uvicorn==0.42.0 ; python_version >= "3.12" and python_version < "4.0"
uvloop==0.21.0 ; python_version >= "3.12" and python_version < "4.0" and sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy"
watchfiles==1.1.0 ; python_version >= "3.12" and python_version < "4.0"
webencodings==0.5.1 ; python_version >= "3.12" and python_version < "4.0"
websockets==15.0.1 ; python_version >= "3.12" and python_version < "4.0"
werkzeug==3.1.6 ; python_version >= "3.12" and python_version < "4.0"
xmlsec==1.3.17 ; python_version >= "3.12" and python_version < "4.0"
zxcvbn==4.5.0 ; python_version >= "3.12" and python_version < "4.0"
